+++ ci4/app/Libraries/Kubernetes/KubeLog.php
@@ -3,7 +3,6 @@
 use App\Libraries\ZMQ\ChangeEvent;
 use App\Libraries\ZMQ\Events;
 use App\Libraries\ZMQ\ZMQProxy;
-use DebugTool\Data;
 use RenokiCo\PhpK8s\Exceptions\KubernetesAPIException;
 use RenokiCo\PhpK8s\Exceptions\KubernetesLogsException;
 use RenokiCo\PhpK8s\Exceptions\KubernetesWatchException;
@@ -23,20 +22,10 @@ class KubeLog {
      */
     public function getLogs(string $namespace, string $podName, string $containerName): array {
         $pod = $this->cluster->getPodByName($podName, $namespace);
-        $logs = explode("\n", $pod->containerLogs($containerName, [
+        return KubeHelper::LogLines($pod->containerLogs($containerName, [
             'tailLines' => 100,
             'timestamps' => true,
         ]));
-        $lines = [];
-        foreach ($logs as $log) {
-            if (strlen($log)) {
-                $lines[] = [
-                    'date' => substr($log, 0, 30),
-                    'line' => substr($log, 31),
-                ];
-            }
-        }
-        return $lines;
     }
 
     /**
@@ -46,15 +35,8 @@ class KubeLog {
     public function watchLog(string $namespace, string $podName, string $containerName): void {
         $pod = $this->cluster->getPodByName($podName, $namespace);
         $pod->watchContainerLogs($containerName, function ($logs) use ($containerName, $podName) {
-            $lines = [];
-            foreach (explode("\n", $logs) as $log) {
-                if (strlen($log)) {
-                    $lines[] = [
-                        'date' => substr($log, 0, 30),
-                        'line' => substr($log, 30),
-                    ];
-                }
-            }
+            // The same split as `getLogs()`, which it used to differ from by one character.
+            $lines = KubeHelper::LogLines($logs);
 
             ZMQProxy::getInstance()->send(
                 Events::KubernetesPod_Logs_Watch($podName, $containerName),